+++ src/componentes/Home.js
@@ -1,7 +1,7 @@
 import React from "react";
-import Header from "./Header";
-import Main from "./Main";
-import Footer from "./Footer";
+import Header from "./Header/Header";
+import Main from "./Main/Main";
+import Footer from "./layout/Footer/Footer";
 
 class Home extends React.Component {
 	state = {
