--- src/components/Home/Home.jsx
@@ -1,17 +1,0 @@
-import React from "react";
-
-class Home extends React.Component {
-	state = {
-		num: 0,
-	};
-
-	render() {
-		return (
-			<div>
-               <p>Hola desde Home</p>
-			</div>
-		);
-	}
-}
-
-export default Home;
