--- src/index.css
@@ -1,11 +1,0 @@
-.App {
-  text-align: center;
-  box-shadow: 0 0.3rem 0.3rem rgba(255, 9, 9, 0.562);
-  padding: 1rem;
-}
-
-button {
-  padding: 1rem;
-  font-size: 1.2rem;
-  background-color: aqua;
-}
