+++ src/index.scss
@@ -1,0 +1,9 @@
+@import "./styles/components/bg-video";
+@import "./styles/components/button";
+@import "./styles/components/defaultFigures";
+
+@import "./styles/utils/animations";
+@import "./styles/utils/grid";
+@import "./styles/utils/utilities";
+
+@import "./styles/styles";
