--- src/styles/abstracts/_extends.scss
@@ -1,11 +1,0 @@
-
-%estructura_bloques {
-    min-width: 100%;
-    max-width: 100%;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    padding: 6rem 0;
-}
