--- src/styles/abstracts/_mixins.scss
@@ -1,81 +1,0 @@
-@import "./variables";
-@import "./extends";
-
-@mixin clearfix {
-    &::after {
-        content: "";
-        display: table;
-        clear: both;
-    }
-}
-
-@mixin absCenter {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-// MEDIA QUERY MANAGER
-
-/*
-
-0 - 600 px:         Phone
-600 - 900px:        Tablet portrait
-900 - 1200px:       Tablet landscape
-[1200 - 1800]       Where normal styles apply (desktop first approach)
-1800px + :          Big desktop
-
-$breakpoint argument choices:
-- phone
-- tab-port
-- tab-land
-- big-desktop
-
-ORDER: base + typography > general layout + grid > page layout > components
-
-1em = 16px
-
-*/
-
-
-@mixin respond($breakpoint) {
-    @if $breakpoint==phone {
-        @media only screen and (max-width: 37.5em) {
-            @content
-        }
-
-        ; //600px
-    }
-
-    @if $breakpoint==tab-port {
-        @media only screen and (max-width: 56.25em) {
-            @content
-        }
-
-        ; //900px
-    }
-
-    @if $breakpoint==tab-land {
-        @media only screen and (max-width: 75em) {
-            @content
-        }
-
-        ; //1200px
-    }
-
-    @if $breakpoint==big-desktop {
-        @media only screen and (min-width: 112.5em) {
-            @content
-        }
-
-        ; //1800
-    }
-}
-
-@mixin estructura_bloque($color_bloque) {
-    @extend %estructura_bloques;
-    background-color: $color_bloque;
-}
-.block_white { @include estructura_bloque($color-plain-white); }
-.block_grey { @include estructura_bloque($color-grey-light-1); }
