--- src/styles/abstracts/_variable.scss
@@ -1,28 +1,0 @@
-//-------- COLORES --------
-$color_resaltado: rgb(235, 187, 45);
-$color_blanco: rgb(255, 255, 255);
-$color_negro: rgb(3, 8, 11);
-$color_gris: rgb(251, 251, 251);
-$color_celeste: rgb(55, 88, 255);
-$color_subrayado: rgb(235, 187, 45);
-$color_video: rgba(0, 18, 34, 0.7);
-$color_oscurecer: rgba(0, 0, 0, 0.25);
-$color_navbar: rgba(255, 255, 255, 0.75);
-$color_navbar_mobile: rgba(0, 18, 34, 0.95);
-$color_footer: rgb(17, 37, 61);
-$color_footer_copyright: rgb(8, 16, 24);
-
-//-------- TAMAÑOS --------
-$icono_ga: 3rem;
-$icono_card: 4rem;
-$icono_redes: 3rem;
-$foto_leader_w: 8rem;
-$foto_leader_h: 7.53125rem;
-$quote: 3rem;
-$cuadradito: 1.5rem;
-
-//-------- MEDIAQUERIES --------
-$tablet: 768px;
-$desktop: 1024px;
-$desktop_m: 1680px;
-$desktop_xl: 1920px;
