--- src/styles/abstracts/_variables.scss
@@ -1,22 +1,0 @@
-// COLORS
-
-$color-primary: #f63372;
-$color-secondary: #fdd300;
-$color-secondary-tertiary: #6aedf7;
-$color-grey-dark: #696969;
-$color-grey-dark-2: #545454;
-$color-grey-light-1: #f4f4f4;
-$color-plain-white: #fff;
-$color-plain-black: #000;
-
-
-// GRID
-
-$grid-width: 114rem;
-$gutter-vertical: 8rem;
-$gutter-vertical-small: 6rem;
-$gutter-horizontal: 6rem;
-
-// FONTS
-
-$default-font-size: 1.6rem;
