--- src/styles/base/_typography.scss
@@ -1,17 +1,0 @@
-body {
-    font-family: "Lato", sans-serif;
-    font-weight: 400;
-    font-size: 1.6rem;
-    line-height: 1.7;
-    color: $color-grey-dark;
-
-    box-sizing: border-box;
-}
-
-.paragraph {
-    font-size: 1.6rem;
-
-    &:not(:last-child) {
-        margin-bottom: 3rem;
-    }
-}
