+++ src/styles/components/_button.scss
@@ -1,3 +1,5 @@
+@import '../styles';
+
 .btn {
 
     &,
@@ -76,7 +78,7 @@
 
     &:link,
     &:visited {
-        font-size: $default-font-size;
+        font-size: $font-h3;
         color: $color-primary;
         display: inline-block;
         text-decoration: none;
