+++ src/styles/components/_defaultFigures.scss
@@ -1,3 +1,5 @@
+@import '../styles';
+
 .cuadrado {
 	width: 15rem;
 	height: 15rem;
