--- src/styles/main.scss
@@ -1,25 +1,0 @@
-@import "abstracts/functions";
-@import "abstracts/mixins";
-@import "abstracts/variables";
-@import "abstracts/extends"; 
-
-@import "base/animations";
-@import "base/base";
-@import "base/typography";
-@import "base/utilities";
-
-
-@import "components/bg-video";
-@import "components/button";
-@import "components/card";
-@import "components/defaultFigures";
-@import "components/feature-box";
-@import "components/popup";
-
-
-@import "layout/grid";
-@import "layout/header";
-@import "layout/footer";
-@import "layout/navigation";
-
-@import "pages/home";
