+++ src/styles/utils/_animations.scss
