+++ src/styles/utils/_utilities.scss
@@ -5,7 +5,7 @@
 .u-margin-bottom-big {
     margin-bottom: 8rem !important;
 
-    @include respond(tab-port) {
+    @include tablet {
         margin-bottom: 5rem !important;
     }
 }
@@ -13,7 +13,7 @@
 .u-margin-bottom-medium {
     margin-bottom: 4rem !important;
 
-    @include respond(tab-port) {
+    @include tablet {
         margin-bottom: 3rem !important;
     }
 }
